minor code style fix

+++ ft_strlcat.c
@@ -6,16 +6,16 @@
 /*   By: csouza-f <csouza-f@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/20 16:05:08 by csouza-f          #+#    #+#             */
-/*   Updated: 2020/02/03 18:03:45 by csouza-f         ###   ########.fr       */
+/*   Updated: 2020/06/10 23:54:54 by caio             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlcat(char *dst, const char *src, unsigned long int dstsize)
+size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 {
-	unsigned long int	i;
-	unsigned long int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	while (i < dstsize && dst[i])
